+++ src/main/java/com/frejdh/util/job/JobFunction.java
@@ -2,12 +2,16 @@ package com.frejdh.util.job;
 
 import com.frejdh.util.job.exceptions.JobAlreadyStartedException;
 import com.frejdh.util.job.model.JobStatus;
-import com.frejdh.util.job.model.interfaces.JobAction;
-import com.frejdh.util.job.model.interfaces.JobCallback;
-import com.frejdh.util.job.model.interfaces.JobError;
-import com.frejdh.util.job.model.interfaces.JobFinalize;
+import com.frejdh.util.job.model.callables.JobAction;
+import com.frejdh.util.job.model.callables.JobOnCallback;
+import com.frejdh.util.job.model.callables.JobOnError;
+import com.frejdh.util.job.model.callables.JobOnFinalize;
+import com.frejdh.util.job.model.callables.JobOnStatusChange;
+import lombok.Builder;
+import lombok.NonNull;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+
 import java.time.Instant;
 
 /**
@@ -16,28 +20,46 @@ import java.time.Instant;
 public class JobFunction {
 
 	private Job job;
-	private final JobAction jobAction;
-	private JobCallback jobCallback;
-	private JobError jobError;
-	private JobFinalize jobFinalize;
+
+	@NonNull private JobAction action;
+	private JobOnCallback onJobCallback;
+	private JobOnError onJobError;
+	private JobOnFinalize onJobFinalize;
+	private JobOnStatusChange onStatusChange;
 	private JobStatus status;
 	private Long startTime;
 	private Long stopTime;
 	private Throwable throwable;
 
-	public JobFunction(JobAction jobAction) {
-		this.jobAction = jobAction;
+	/**
+	 * Lombok builder alternative (only uses fields in constructor)
+	 */
+	@Builder
+	protected JobFunction(@NonNull JobAction action,
+						  JobOnCallback onJobCallback,
+						  JobOnError onJobError,
+						  JobOnFinalize onJobFinalize,
+						  JobOnStatusChange onStatusChange) {
+		this.action = action;
+		this.onJobCallback = onJobCallback;
+		this.onJobError = onJobError;
+		this.onJobFinalize = onJobFinalize;
+		this.onStatusChange = onStatusChange;
 	}
 
-	public JobFunction(@NotNull JobAction action,
-					   @Nullable JobCallback onSuccess,
-					   @Nullable JobError onError,
-					   @Nullable JobFinalize onComplete) {
-		this.jobAction = action;
-		this.jobCallback = onSuccess;
-		this.jobError = onError;
-		this.jobFinalize = onComplete;
+	public JobFunction(@NonNull JobAction action) {
+		this.action = action;
 		this.status = JobStatus.INITIALIZED;
+	}
+
+	public JobFunction(@NonNull JobAction action,
+					   @Nullable JobOnCallback onSuccess,
+					   @Nullable JobOnError onError,
+					   @Nullable JobOnFinalize onComplete) {
+		this(action);
+		this.onJobCallback = onSuccess;
+		this.onJobError = onError;
+		this.onJobFinalize = onComplete;
 	}
 
 	JobFunction setJob(Job job) {
@@ -46,22 +68,35 @@ public class JobFunction {
 	}
 
 	void setStatus(JobStatus status) {
+		final boolean isStatusChanged = this.status == null || !this.status.equals(status);
 		this.status = status;
+		if (onStatusChange != null && isStatusChanged) {
+			onStatusChange.onStatusChange();
+		}
 	}
 
-	public JobFunction setCallback(JobCallback jobCallback) {
-		this.jobCallback = jobCallback;
-		return this;
+	void setAction(@NotNull JobAction jobAction) {
+		this.action = jobAction;
 	}
 
-	public JobFunction setOnError(JobError jobError) {
-		this.jobError = jobError;
-		return this;
+	void setOnCallback(JobOnCallback jobCallback) {
+		this.onJobCallback = jobCallback;
 	}
 
-	public JobFunction setFinalize(JobFinalize jobFinalize) {
-		this.jobFinalize = jobFinalize;
-		return this;
+	void setOnError(JobOnError jobError) {
+		this.onJobError = jobError;
+	}
+
+	void setOnFinalize(JobOnFinalize jobFinalize) {
+		this.onJobFinalize = jobFinalize;
+	}
+
+	void setOnStatusChange(JobOnStatusChange onStatusChange) {
+		this.onStatusChange = onStatusChange;
+	}
+
+	JobOnError getJobOnError() {
+		return this.onJobError;
 	}
 
 	public Job getJob() {
@@ -86,30 +121,31 @@ public class JobFunction {
 	 * Start the job function (and callback)
 	 */
 	public void start() throws JobAlreadyStartedException {
+		checkIfStartedAlready();
 		synchronized (this) {
 			checkIfStartedAlready();
 			this.startTime = Instant.now().toEpochMilli();
 			try {
-				status = JobStatus.RUNNING_ACTION;
-				jobAction.action();
+				setStatus(JobStatus.RUNNING_ACTION);
+				action.action();
 
-				if (jobCallback != null) {
-					status = JobStatus.RUNNING_CALLBACK;
-					jobCallback.callback(job);
+				if (onJobCallback != null) {
+					setStatus(JobStatus.RUNNING_CALLBACK);
+					onJobCallback.callback(job);
 				}
 
-				status = JobStatus.FINISHED;
+				setStatus(JobStatus.FINISHED);
 			} catch (Throwable throwable) {
-				status = JobStatus.FAILED;
+				setStatus(JobStatus.FAILED);
 				this.throwable = throwable;
-				if (jobError != null) {
-					jobError.onError(throwable);
+				if (onJobError != null) {
+					onJobError.onError(throwable);
 				}
 			} finally {
 				stopTime = Instant.now().toEpochMilli();
-				if (jobFinalize != null) {
+				if (onJobFinalize != null) {
 					try {
-						jobFinalize.onComplete();
+						onJobFinalize.onComplete();
 					} catch (Exception e) {
 						e.printStackTrace();
 					}