+++ src/main/java/com/frejdh/util/job/model/QueueOptions.java
@@ -1,23 +1,35 @@
 package com.frejdh.util.job.model;
 
 import com.frejdh.util.job.Job;
+import com.frejdh.util.job.model.callables.JobOnError;
 import lombok.Builder;
 import lombok.Getter;
+
+import java.util.ArrayList;
 import java.util.List;
 
+@SuppressWarnings("FieldMayBeFinal")
 @Builder(toBuilder = true, setterPrefix = "with")
 @Getter
 public class QueueOptions {
 	public static final int DEFAULT_MAX_PARALLEL_JOBS = 4;
 	public static final int DEFAULT_AMOUNT_OF_THREADS = 1;
 
+	@Builder.Default
 	private int maxParallelJobs = DEFAULT_MAX_PARALLEL_JOBS;
-	private int amountOfThreads = DEFAULT_AMOUNT_OF_THREADS;
-	private String persistenceFile;
-	private boolean singleExecution;
-	private List<Job> predefinedJobs;
 
-	public boolean hasPersistenceFile() {
-		return persistenceFile != null;
+	@Builder.Default
+	private int amountOfThreads = DEFAULT_AMOUNT_OF_THREADS;
+
+	@Builder.Default
+	private boolean singleExecution = false;
+
+	@Builder.Default
+	private List<Job> predefinedJobs = new ArrayList<>();
+
+	private JobOnError onJobError;
+
+	public static QueueOptions getDefault() {
+		return QueueOptions.builder().build();
 	}
 }