+++ src/main/java/com/frejdh/util/job/model/callables/JobOnCallback.java
@@ -1,8 +1,8 @@
-package com.frejdh.util.job.model.interfaces;
+package com.frejdh.util.job.model.callables;
 
 import com.frejdh.util.job.Job;
 
 @FunctionalInterface
-public interface JobCallback {
+public interface JobOnCallback {
 	void callback(Job job);
 }