+++ src/main/java/com/frejdh/util/job/persistence/impl/h2/H2DaoService.java
@@ -2,10 +2,32 @@ package com.frejdh.util.job.persistence.impl.h2;
 
 import com.frejdh.util.job.Job;
 import com.frejdh.util.job.persistence.AbstractDaoService;
+import org.jetbrains.annotations.NotNull;
+
 import java.util.List;
 import java.util.Map;
 
 public class H2DaoService extends AbstractDaoService {
+
+	@Override
+	public Job addJob(@NotNull Job job) {
+		return null;
+	}
+
+	@Override
+	public Job updateJob(@NotNull Job job) {
+		return null;
+	}
+
+	@Override
+	public Job updateJobOnlyOnFreeResource(@NotNull Job job) {
+		return null;
+	}
+
+	@Override
+	protected Job removeJob(@NotNull Job job) {
+		return null;
+	}
 
 	@Override
 	public Job getJobById(Long id) {
@@ -13,7 +35,7 @@ public class H2DaoService extends AbstractDaoService {
 	}
 
 	@Override
-	public Map<Long, Job> getPendingJobsById() {
+	public Map<Long, Job> getPendingJobs() {
 		return null;
 	}
 
@@ -33,12 +55,12 @@ public class H2DaoService extends AbstractDaoService {
 	}
 
 	@Override
-	public Map<Long, Job> getRunningJobsById() {
+	public Map<Long, Job> getRunningJobs() {
 		return null;
 	}
 
 	@Override
-	public Map<String, Job> getRunningJobsByResource() {
+	public Map<String, Job> getRunningJobsForResources() {
 		return null;
 	}
 
@@ -54,6 +76,16 @@ public class H2DaoService extends AbstractDaoService {
 
 	@Override
 	public List<Job> getFinishedJobsByResource(String resource) {
+		return null;
+	}
+
+	@Override
+	public Job getLastAddedJob() {
+		return null;
+	}
+
+	@Override
+	public Job getLastFinishedJob() {
 		return null;
 	}
 }