+++ src/main/java/com/frejdh/util/job/state/LocalJobWorkerThreadState.java
@@ -2,6 +2,8 @@ package com.frejdh.util.job.state;
 
 import com.frejdh.util.job.Job;
 import java.util.concurrent.Future;
+import java.util.concurrent.atomic.AtomicBoolean;
+
 
 public class LocalJobWorkerThreadState {
 	public Future<?> jobExecutionFuture;