+++ changelog.md
@@ -3,8 +3,12 @@ Changelog
 
 # 1.1.0
 - Stoppable jobs.
-- Configurations on a job-basis.
-- Persistence option
+- Configurations per job-basis.
+  - Timeout option is currently the only added one (so far).
+- Persistence configurations available for jobs.
+    - Optional H2 file implementation added.
+- Moved to strict builder usage for classes (`Job` / `JobFunction`, etc).
+
 
 
 # 1.0.0